Support fractional item quantities (split kits)
Upstream orders carry fractional quantities (0.5 of a split kit) — the
integer qty column made every such order's items import fail since day
one. Migration 1786468400 (applied): order_items.qty becomes NUMERIC(10,2)
with the dependent view chain (v_moq_progress → v_product_profit →
v_vendor_balances / v_group_buy_pnl) dropped and recreated around the
ALTER at their current definitions. upsertOrderItem/deleteOrderItemsNotIn
cast ::numeric; the paste parser accepts decimal quantities. Two 0.5
orders sum to 1 unit of vendor demand in the views, which is exactly the
split-kit economics.

+++ src/actions/orders/deleteOrderItemsNotIn.ts
@@ -13,7 +13,7 @@ function deleteOrderItemsNotIn() {
       WHERE oi.order_id = {{params.order_id}}::bigint
         AND oi.group_buy_product_id NOT IN (
           SELECT gbp.id
-          FROM jsonb_to_recordset({{params.items}}::jsonb) AS x(sku text, qty int)
+          FROM jsonb_to_recordset({{params.items}}::jsonb) AS x(sku text, qty numeric)
           JOIN products p ON p.sku_code = x.sku
           JOIN group_buy_products gbp ON gbp.product_id = p.id
             AND gbp.group_buy_id = {{params.group_buy_id}}::bigint
